[Graphics] 補上漏掉的 function

+++ Source/GraphicKernel/VertexDescription.h
@@ -83,7 +83,7 @@ namespace Enigma::Graphics
     public:
         enum class ColorType
         {
-            Byte,
+            UInt,
             Float
         };
     public:
